CX-9118: Alert Documentation added
